Add files via upload
This code purpose to create a list of stop words from a single text input. Stop words are the expressions irrelevants for text analises, such as articles, pronouns and others that should be removed from the text before creating a counted word list that can be used to find the subject and other information regarding the words presented in the text inserted.  Obviously, there are already many lists for this expressions, and many ways to generate more accurate selections of words that are repeated constantly in a text, therefore don`t contribute for a further analysis of what is told. Also, is very necessary take care to clean the list of possible words that even repeated several times are the ones that offer the results for the research in progress. The line 31 contains the quantity of occurences that are consider to denote possible non relevant word. Besides the manual overview check, this number can be changed to accurate the results.