Add files via upload
This is a single code algorithm that create a word cloud from Python generated data in a HTML + CSS styling code. Python's string format allows to use iterable lists to fill gaps with frequency and corresponding word for a given text. The code also ask for higher and lower frequency limits to exclude words irrelevant to the analysis.